Rename files, refactored tree
