+++ Data/Antenna/Availability/Nominal/NominalAntennaAvailability.csv
