+++ Images/TransitionTrajectoryPlot.png
